Add validations and error handlers into Http requests

+++ src/main/java/com/example/cashcard/CashCardController.java
@@ -40,7 +40,7 @@ public class CashCardController {
 	public ResponseEntity<String> rootPath() {
 		String response = """
 				{
-					"message": "Welcome to cash cards APP"
+					"message": "Welcome to cash cards API REST"
 				}
 				""";
 
@@ -50,15 +50,13 @@ public class CashCardController {
 	@GetMapping("/all")
 	public ResponseEntity<List<CashCard>> findAll(Pageable pageable) {
 		Page<CashCard> page = cashCardRepository.findAll(
-			PageRequest.of(
-				pageable.getPageNumber(), 
-				pageable.getPageSize(), 
-				pageable.getSortOr(Sort.by(Sort.Direction.ASC, "amount"))
-				)
-		);
+				PageRequest.of(
+						pageable.getPageNumber(),
+						pageable.getPageSize(),
+						pageable.getSortOr(Sort.by(Sort.Direction.ASC, "amount"))));
 
 		return ResponseEntity.ok(page.getContent());
-		//return ResponseEntity.ok(cashCardRepository.findAll());
+		// return ResponseEntity.ok(cashCardRepository.findAll());
 	}
 
 	@GetMapping("/{requestedId}")
@@ -75,12 +73,18 @@ public class CashCardController {
 	private ResponseEntity<Void> AddNewCashCard(@RequestParam String username, @RequestParam Double amount,
 			UriComponentsBuilder ucb) {
 		try {
-			// faltan validaciones de si ya existe ese ID
 			CashCard newCashCard = new CashCard();
-			newCashCard.setId((int) (Math.random() * 100000 + 1));
+			Optional<CashCard> cashCardSearched;
+			Integer newId;
+			do {
+				newId = (int) (Math.random() * 10000000 + 1);
+				cashCardSearched = cashCardRepository.findById(newId);
+			} while (cashCardSearched.isPresent());
+
+			// newCashCard.setId((int) (Math.random() * 100000 + 1));
+			newCashCard.setId(newId);
 			newCashCard.setUsername(username);
 			newCashCard.setAmount(amount);
-			System.out.println("new ID: " + newCashCard.getId());
 			CashCard savedCashCard = cashCardRepository.save(newCashCard);
 			URI locationOfNewCashCard = ucb.path("/cashcards/{id}").buildAndExpand(savedCashCard.getId()).toUri();
 			return ResponseEntity.created(locationOfNewCashCard).build();
@@ -93,14 +97,20 @@ public class CashCardController {
 	// PUT
 	@PutMapping("/{requestedId}")
 	public ResponseEntity<Void> putCashCard(@PathVariable Integer requestedId, @RequestBody CashCard cashCardUpdate) {
-		// Faltan validaciones
 		Optional<CashCard> cashCard = cashCardRepository.findById(requestedId);
-		CashCard updatedCashCard = new CashCard();
-		updatedCashCard.setId(cashCard.get().getId());
-		updatedCashCard.setUsername(cashCardUpdate.getUsername());
-		updatedCashCard.setAmount(cashCardUpdate.getAmount());
-		cashCardRepository.save(updatedCashCard);
-		return ResponseEntity.noContent().build();
+		if (cashCard.isPresent()) {
+			CashCard updatedCashCard = new CashCard();
+			updatedCashCard.setId(cashCard.get().getId());
+			updatedCashCard.setUsername(cashCardUpdate.getUsername());
+			updatedCashCard.setAmount(cashCardUpdate.getAmount());
+			try {
+				cashCardRepository.save(updatedCashCard);
+				return ResponseEntity.noContent().build();
+			} catch (Exception e) {
+				return ResponseEntity.badRequest().build();
+			}
+		}
+		return ResponseEntity.notFound().build();
 	}
 
 	// PATCH
@@ -109,13 +119,17 @@ public class CashCardController {
 			@RequestBody Map<String, Object> fields) {
 		Optional<CashCard> cashCard = cashCardRepository.findById(requestedId);
 		if (cashCard.isPresent()) {
-			fields.forEach((key, value) -> {
-				Field field = ReflectionUtils.findField(CashCard.class, (String) key);
-				field.setAccessible(true);
-				ReflectionUtils.setField(field, cashCard.get(), value);
-			});
-			cashCardRepository.save(cashCard.get());
-			return ResponseEntity.ok(cashCard.get());
+			try {
+				fields.forEach((key, value) -> {
+					Field field = ReflectionUtils.findField(CashCard.class, (String) key);
+					field.setAccessible(true);
+					ReflectionUtils.setField(field, cashCard.get(), value);
+				});
+				cashCardRepository.save(cashCard.get());
+				return ResponseEntity.ok(cashCard.get());
+			} catch (Exception e) {
+				return ResponseEntity.badRequest().build();
+			}
 
 		} else {
 			return ResponseEntity.notFound().build();
@@ -125,8 +139,13 @@ public class CashCardController {
 	// DELETE
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> deleteCashCard(@PathVariable Integer id) {
-		cashCardRepository.deleteById(id);
-		return ResponseEntity.noContent().build();
+		Optional<CashCard> cashCard = cashCardRepository.findById(id);
+		if (cashCard.isPresent()) {
+			cashCardRepository.deleteById(id);
+			return ResponseEntity.noContent().build();
+		} else {
+			return ResponseEntity.notFound().build();
+		}
 
 	}
